+++ kirlent_docutils/__init__.py
@@ -9,12 +9,17 @@ __version__ = "0.1.2"
 
 from docutils.core import publish_cmdline
 
-from . import html5, impressjs, revealjs
+from . import html5, impressjs, revealjs, slides
 
 
 def publish_cmdline_html5(*args, **kwargs):
     """Run utility for converting an RST file to HTML5."""
     publish_cmdline(*args, **kwargs, writer=html5.Writer())
+
+
+def publish_cmdline_slides(*args, **kwargs):
+    """Run utility for converting an RST file to an HTML5 presentation."""
+    publish_cmdline(*args, **kwargs, writer=slides.Writer())
 
 
 def publish_cmdline_impressjs(*args, **kwargs):